test(radio-group): migrate tests to playwright (#25713)
